scraper/requirements.txt
requests==2.25.1
psycopg2==2.9.1
pytz==2021.1
python-dotenv==0.19.2